+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New API `ReadOnlyColorSchemeWrapper`. It just like `ReadOnlyXxxWrapper` classes in JavaFX.
 - New API `ColorStyle#isSupported(ColorSpecVersion)` that returns the support status of a color style under 
   a specific color specification version.
+- New API `ColorScheme#extractColor(Image)` for extracting representative colors from images.
 
 ### Breaking changes
 
@@ -19,7 +20,7 @@
 
 ### Theme Builder
 
-- Improved support for the 2023 version of the Material 3 color system specification and the watch platform in MonetFX Theme Builder.
+- Improved support for the 2025 version of the Material 3 color system specification and the watch platform in MonetFX Theme Builder.
 
   Now, when switching the target platform to watch, the 2025 specification will be automatically selected and dark mode enabled;